chore: restructure action test folders
