+++ client/__test__/actions/createGroup.test.jsx
@@ -1,5 +1,5 @@
 import expect from 'expect';
-import createGroup from '../../actions/createGroup';
+import createGroup from '../../actions/groupActions';
 
 describe('Create Group Action', () => {
   it('should contain createGroup function', () => {
