+++ client/__test__/actions/getMessages.test.jsx
@@ -1,7 +1,7 @@
 import expect from 'expect';
 import thunk from 'redux-thunk';
 import configureMockStore from 'redux-mock-store';
-import getMessages, { setMessages } from '../../actions/getMessages';
+import getMessages, { setMessages } from '../../actions/messageActions';
 import * as types from '../../actions/types';
 
 const middlewares = [thunk];
