+++ client/__test__/actions/passmessageAction.test.jsx
@@ -1,5 +1,5 @@
 import expect from 'expect';
-import passMessage from '../../actions/passMessageAction';
+import { passMessage } from '../../actions/messageActions';
 
 describe('Pass message Action', () => {
   it('should contain passMessage object', () => {
