+++ client/__test__/actions/removeUserAction.test.jsx
@@ -1,5 +1,5 @@
 import expect from 'expect';
-import removeUser from '../../actions/removeUserAction';
+import { removeUser } from '../../actions/groupActions';
 
 describe('Remove User Action', () => {
   it('should return a function', () => {
