--- client/actions/passMessageAction.js
@@ -1,16 +1,0 @@
-/**
- * action to pass a message to store
- */
-
-import { PASS_MESSAGE } from './types';
-
-/**
- * @return {object} returns object and action type
- * @param {object} data object containing message detail
- */
-const passMessage = data => ({
-  type: PASS_MESSAGE,
-  data
-});
-
-export default passMessage;
