--- client/actions/postMessageAction.js
@@ -1,15 +1,0 @@
-/**
- * action to post message in a group
- */
-
-import axios from 'axios';
-
-/**
- * @return {promise} returns an array with server response
- * @param {number} id
- * @param {object} data
- */
-const postMessage = (id, data) =>
-  () => axios.post(`/api/v1/group/${id}/message`, data);
-
-export default postMessage;
