+++ client/index.jsx
@@ -9,10 +9,10 @@ import { Provider } from 'react-redux';
 import jwt from 'jsonwebtoken';
 import store from './store';
 import setAuthToken from './utilities/setAuthToken';
-import App from './components/App';
+import App from './components/App.jsx';
 import { setCurrentUser } from './actions/signinAction';
-import setGroupId from './actions/groupIdAction';
-import passMessage from './actions/passMessageAction';
+import { setGroupId } from './actions/groupActions';
+import { passMessage } from './actions/messageActions';
 import './styles/main.scss';
